- fix integer overflow causes minus

+++ app/src/main/java/lab/justonebyte/simpleexpense/ui/components/CustomPieChartWithData.kt
@@ -72,7 +72,7 @@ class MyValueFormatter : ValueFormatter() {
 val colors = listOf(
     bar1,bar2,bar3,bar4,bar5,bar6,bar7,bar8,bar9,bar10,bar11,bar12,bar13,bar14,bar15,bar16,bar17,bar18,bar19,bar20
 )
-fun calculatePercentage(part: Int, whole: Int): Double {
+fun calculatePercentage(part: Long, whole: Long): Double {
     return (part.toDouble() / whole.toDouble()) * 100
 }
 
@@ -84,8 +84,8 @@ fun CustomPieChartWithData(
     transactionType: TransactionType
 ){
     val materialColor =  MaterialTheme.colorScheme.primary.toArgb()
-    val totalAmount = transactions.sumOf { it.amount }
-    val pieEntries = transactions.groupBy { it.category }.map { PieEntry(calculatePercentage(it.value.sumOf { s-> s.amount },totalAmount).toFloat(),it.key.name) }.sortedByDescending { it.value }
+    val totalAmount = transactions.sumOf { it.amount.toLong() }
+    val pieEntries = transactions.groupBy { it.category }.map { PieEntry(calculatePercentage(it.value.sumOf { s-> s.amount.toLong() },totalAmount).toFloat(),it.key.name) }.sortedByDescending { it.value }
 
     if(pieEntries.isNotEmpty()){
         Column(
